Support ItemTemplate in MultiComboBox (#233)
`ItemTemplate` is a regular functionnality of `ItemsControl`, but I didn't support it correctly in the original implementation of `MultiComboBox` due to forgetting to bind the `ContentTemplate` properly on the `ContentPresenter` of the items.
I also added a little bit of logic to use the `ItemTemplate` in the SelectedItem's template when a specific `SelectedItemTemplate` is not provided, so that by default, if you specify an `ItemTemplate`, the `ClosableTag`s will display the same template as the items unless overriden.